adding locking to customer resource manager

+++ src/ResImpl/CustomerResourceManager.java
@@ -3,17 +3,21 @@ package ResImpl;
 import java.rmi.RemoteException;
 import java.util.Calendar;
 import java.util.Enumeration;
+import java.util.Stack;
 import java.util.Vector;
 
 import LockManager.DeadlockException;
+import LockManager.TrxnObj;
 import ResInterface.CarBackend;
 import ResInterface.CustomerFrontend;
 import ResInterface.FlightBackend;
 import ResInterface.ReservationFrontend;
 import ResInterface.RoomBackend;
+import Transactions.ITransactionManager;
 import Transactions.InvalidTransactionException;
+import Transactions.Operation;
 
-public class CustomerResourceManager extends AbstractResourceManager implements CustomerFrontend,ReservationFrontend {
+public class CustomerResourceManager extends AbstractResourceManager implements ITransactionManager, CustomerFrontend,ReservationFrontend {
 	private FlightBackend flightRM;
 	private RoomBackend roomRM;
 	private CarBackend carRM;
@@ -28,27 +32,43 @@ public class CustomerResourceManager extends AbstractResourceManager implements 
 	}
 	
 	@Override
-	public int newCustomer(int id) throws DeadlockException {
+	public int newCustomer(int id) throws DeadlockException, InvalidTransactionException {
 		Trace.info("INFO: RM::newCustomer(" + id + ") called" );
+		Stack<Operation> ops = activeTransactions.get(id);
+		if (ops == null){
+			throw new InvalidTransactionException("No transaction with id "+id);
+		}
+		
 		// Generate a globally unique ID for the new customer
 		int cid = Integer.parseInt( String.valueOf(id) +
 								String.valueOf(Calendar.getInstance().get(Calendar.MILLISECOND)) +
 								String.valueOf( Math.round( Math.random() * 100 + 1 )));
 		Customer cust = new Customer( cid );
+		
+		lockManager.Lock(id, cust.getKey(), TrxnObj.WRITE);
 		writeData( id, cust.getKey(), cust );
 		Trace.info("RM::newCustomer(" + cid + ") returns ID=" + cid );
 		return cid;
 	}
 
 	@Override
-	public boolean newCustomer(int id, int customerID ) throws DeadlockException {
+	public boolean newCustomer(int id, int customerID ) throws DeadlockException, InvalidTransactionException {
 		Trace.info("INFO: RM::newCustomer(" + id + ", " + customerID + ") called" );
+		Stack<Operation> ops = activeTransactions.get(id);
+		if (ops == null){
+			throw new InvalidTransactionException("No transaction with id "+id);
+		}
+		
 		Customer cust = (Customer) readData( id, Customer.getKey(customerID) );
 		if( cust == null ) {
-			cust = new Customer(customerID);
-			writeData( id, cust.getKey(), cust );
-			Trace.info("INFO: RM::newCustomer(" + id + ", " + customerID + ") created a new customer" );
-			return true;
+			if (lockManager.Lock(id, Customer.getKey(customerID), TrxnObj.WRITE)){
+				cust = new Customer(customerID);
+				writeData( id, cust.getKey(), cust );
+				Trace.info("INFO: RM::newCustomer(" + id + ", " + customerID + ") created a new customer" );
+				return true;
+			} else {
+				return false;
+			}
 		} else {
 			Trace.info("INFO: RM::newCustomer(" + id + ", " + customerID + ") failed--customer already exists");
 			return false;
@@ -56,8 +76,13 @@ public class CustomerResourceManager extends AbstractResourceManager implements 
 	}
 
 	@Override
-	public boolean deleteCustomer(int id, int customerID) throws DeadlockException {
+	public boolean deleteCustomer(int id, int customerID) throws DeadlockException, InvalidTransactionException {
 		Trace.info("RM::deleteCustomer(" + id + ", " + customerID + ") called" );
+		Stack<Operation> ops = activeTransactions.get(id);
+		if (ops == null){
+			throw new InvalidTransactionException("No transaction with id "+id);
+		}
+		
 		Customer cust = (Customer) readData( id, Customer.getKey(customerID) );
 		if( cust == null ) {
 			Trace.warn("RM::deleteCustomer(" + id + ", " + customerID + ") failed--customer doesn't exist" );
@@ -99,33 +124,58 @@ public class CustomerResourceManager extends AbstractResourceManager implements 
 				
 			}
 			
-			// remove the customer from the storage
-			removeData(id, cust.getKey());
-			
-			Trace.info("RM::deleteCustomer(" + id + ", " + customerID + ") succeeded" );
-			return true;
+			if (lockManager.Lock(id, cust.getKey(), TrxnObj.WRITE)){
+				// remove the customer from the storage
+				removeData(id, cust.getKey());
+				
+				Trace.info("RM::deleteCustomer(" + id + ", " + customerID + ") succeeded" );
+				return true;
+			} else {
+				return false;
+			}
 		}
 	}
 
 	@Override
-	public String queryCustomerInfo(int id, int customerID) throws DeadlockException {
+	public String queryCustomerInfo(int id, int customerID) throws DeadlockException, InvalidTransactionException {
 		Trace.info("RM::queryCustomerInfo(" + id + ", " + customerID + ") called" );
+		
+		Stack<Operation> ops = activeTransactions.get(id);
+		if (ops == null){
+			throw new InvalidTransactionException("No transaction with id "+id);
+		}
+		
 		Customer cust = (Customer) readData( id, Customer.getKey(customerID) );
 		if( cust == null ) {
 			Trace.warn("RM::queryCustomerInfo(" + id + ", " + customerID + ") failed--customer doesn't exist" );
 			return "";   // NOTE: don't change this--WC counts on this value indicating a customer does not exist...
 		} else {
+			if (lockManager.Lock(id, cust.getKey(), TrxnObj.READ)){
 				String s = cust.printBill();
 				Trace.info("RM::queryCustomerInfo(" + id + ", " + customerID + "), bill follows..." );
 				System.out.println( s );
 				return s;
+			} else {
+				return "";
+			}
 		} 
 	}
 
 	@Override
 	public boolean reserveCar(int id, int customer, String location) throws RemoteException, InvalidTransactionException, DeadlockException {
+		Stack<Operation> ops = activeTransactions.get(id);
+		if (ops == null){
+			throw new InvalidTransactionException("No transaction with id "+id);
+		}
+		
 		Car car = carRM.getCar(id, location);
-		boolean success = this.reserveItem(id, customer, car, location);
+		
+		
+		boolean success = false;
+		if (lockManager.Lock(id, Customer.getKey(customer), TrxnObj.WRITE)){
+			success = this.reserveItem(id, customer, car, location);
+		} 
+		
 		if (success){
 			// send back
 			carRM.updateCar(id, location, car);
@@ -136,8 +186,17 @@ public class CustomerResourceManager extends AbstractResourceManager implements 
 	@Override
 	public boolean reserveFlight(int id, int customer, int flightNumber)
 			throws RemoteException, DeadlockException, InvalidTransactionException {
+		Stack<Operation> ops = activeTransactions.get(id);
+		if (ops == null){
+			throw new InvalidTransactionException("No transaction with id "+id);
+		}
+		
 		Flight flight = flightRM.getFlight(id, flightNumber);
-		boolean success = reserveItem(id, customer, flight, String.valueOf(flightNumber));
+		boolean success = false;
+		if (lockManager.Lock(id, Customer.getKey(customer), TrxnObj.WRITE)){
+			success = reserveItem(id, customer, flight, String.valueOf(flightNumber));
+		}
+		
 		if (success)
 			flightRM.updateFlight(id, flightNumber, flight);
 		return success;
@@ -147,7 +206,10 @@ public class CustomerResourceManager extends AbstractResourceManager implements 
 	public boolean reserveRoom(int id, int customer, String location)
 			throws RemoteException, DeadlockException, InvalidTransactionException {
 		Hotel room = roomRM.getRoom(id, location);
-		boolean success = reserveItem(id, customer, room, location);
+		boolean success = false;
+		if (lockManager.Lock(id, Customer.getKey(customer), TrxnObj.WRITE)){
+			success = reserveItem(id, customer, room, location);
+		}
 		if (success)
 			roomRM.updateRoom(id, location, room);
 		return success;
